backend/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
motor==3.4.0
pydantic==2.7.1
